Generate a request for the Sheets v4 API

Generate a request, using some knowledge of the Sheets v4API. Most users should, instead, use higher-level wrappers that facilitate common tasks, such as reading the contents of a sheet. The functions here are intended for internal use and for programming around the Sheets API.

gs_generate_request(endpoint = character(), params = list(),
  .api_key = api_key())

gs_build_request(path, method, params = list(), .api_key = api_key())

Arguments

endpoint
character(1) Nickname for one of the documented Sheets v4 API endpoints. to do: list or link, once I've auto-generated those docs

params
named list() Parameters destined for endpoint URL substitution or, otherwise, the query.

.api_key
character(1) A valid APIkey. Defaults to a key built into the googlesheets package.

path
character, e.g., "v4/spreadsheets/{spreadsheetId}/values/{range}:append". It can include variables inside curly brackets, as the example does, which are substituted using named parameters found in the params argument.

method
character, should match an HTTP verb, e.g., GET, POST, or PUT

Value

list() Components are method, path, query, and url, suitable as input for gs_make_request(). The path is post-substitution and the query is a named list of all the input params that were not used during this substitution. url is the full URL after prepending the base URL for the Sheets v4 API and appending an API key to the query.

Details

There are two functions:

gs_generate_request() is a wrapper around gs_build_request(). It takes a nickname for an endpoint and uses the API spec to look up the path and method. The params are checked for validity and completeness with respect to the endpoint. It then passes things along to gs_build_request().

gs_build_request() builds a request from explicit parts. It is quite dumb, only doing URL endpoint substitution and URL formation. It's up to the caller to make sure the path, method, and params are valid.

Examples

req <- gs_generate_request(
  "spreadsheets.values.append",
  list(
    spreadsheetId = "abc",
    range = "Sheet1!A1:B2",
    responseValueRenderOption = "FORMULA"
  )
)
req

#> $method
#> [1] "POST"
#> 
#> $path
#> v4/spreadsheets/abc/values/Sheet1!A1:B2:append
#> 
#> $query
#> $query$responseValueRenderOption
#> [1] "FORMULA"
#> 
#> $query$key
#> [1] "AIzaSyBTBKXVfmtjOSYmn7wm6S4yMvqdeBMtZMI"
#> 
#> 
#> $url
#> [1] "https://sheets.googleapis.com/v4/spreadsheets/abc/values/Sheet1!A1:B2:append?responseValueRenderOption=FORMULA&key=AIzaSyBTBKXVfmtjOSYmn7wm6S4yMvqdeBMtZMI"
#> 

req <- gs_build_request(
  path = "v4/spreadsheets/{spreadsheetId}/values/{range}:append",
  method = "POST",
  list(
    spreadsheetId = "abc",
    range = "Sheet1!A1:B2",
    responseValueRenderOption = "FORMULA"
  )
)
req

#> $method
#> [1] "POST"
#> 
#> $path
#> v4/spreadsheets/abc/values/Sheet1!A1:B2:append
#> 
#> $query
#> $query$responseValueRenderOption
#> [1] "FORMULA"
#> 
#> $query$key
#> [1] "AIzaSyBTBKXVfmtjOSYmn7wm6S4yMvqdeBMtZMI"
#> 
#> 
#> $url
#> [1] "https://sheets.googleapis.com/v4/spreadsheets/abc/values/Sheet1!A1:B2:append?responseValueRenderOption=FORMULA&key=AIzaSyBTBKXVfmtjOSYmn7wm6S4yMvqdeBMtZMI"
#>